[#438] cleanup
