+++ src/main/java/de/bonndan/nivio/output/map/svg/GroupAreaStyle.java
@@ -1,0 +1,8 @@
+package de.bonndan.nivio.output.map.svg;
+
+public enum GroupAreaStyle {
+    SHARP,
+    WOBBLY,
+    SMOOTH,
+    HEXES
+}
